move App.tsx to root, reconfigure entry point, and add android package to prepare for expo eject
