+++ package.json
@@ -2,7 +2,7 @@
   "name": "nightlight-frontend",
   "author": "nightlight",
   "version": "1.0.0",
-  "main": "node_modules/expo/AppEntry.js",
+  "main": "./App.tsx",
   "keywords": [
     "react",
     "expo",
